odb_loose/indexer: write objects/indexes with 0444 file permissions
Prior to this change, new loose objects were written to disk with a
file mode of 0600. This is in contrast to Git, which writes out
objects with a file mode of 0444.

This was happening because odb_loose was using filebuf's
GIT_FILEBUF_TEMPORARY flag, which creates the buffer file using
mkstemp(3). On most systems, mkstemp(3) will create temporary files
with 0600. It then called git_filebuf_commit_at() to finalize object
creation, which simply moves the temporary buffer file to its
destination.

To address this issue, git_filebuf_commit_at() now requires the
destination file mode to be specified. It then chmods the file before
renaming it to its destination path. Currently the only user of
git_filebuf_commit_at() is odb_loose.

Tests have been added to confirm that new commit, tag, and tree
objects are created with the right permissions. I don't have access to
Windows, so for now I've guarded the tests with "#ifndef GIT_WIN32".

Note: This patch was originally for 0.14 and only affected
odb_loose. As of 0.15 it also makes the indexer write out idx files
with 0444 permissions (like Git), and it also makes transport_git
write out (temporary?) pack files with 0444 permissions.