odb_pack: ignore missing .pack files when refreshing all packs
For reasons I don't yet understand, it's sometimes possible to end up
with a Git repo that has an index for a pack but no corresponding pack
object. This might happen when pushing up fast forwards and making
entire sets of packed objects unreachable.

In any event, when stock Git finds an .idx file with no corresponding
.pack file, it silently ignores them. It only reports this as a
problem when using verify-pack.

To match Git's behavior, this change adds a new GIT_EMISSINGPACKFILE
error. It's currently only used internally by odb_pack. When it sees
the error, it skips over the pack file. If there's ever a public API
for verifying pack files, it can let this error bubble up.